+++ sourced/ml/__init__.py
@@ -1,8 +1,4 @@
 import modelforge.configuration
 
-from sourced.ml.df import DocumentFrequencies
-from sourced.ml.id2vec import Id2Vec
-from sourced.ml.topics import Topics
-
 __version__ = 0, 4, 0
 modelforge.configuration.refresh()
