+++ sourced/ml/models/__init__.py
@@ -1,0 +1,5 @@
+from sourced.ml.models.bow import BOW, NBOW
+from sourced.ml.models.coocc import Cooccurrences
+from sourced.ml.models.df import DocumentFrequencies
+from sourced.ml.models.id2vec import Id2Vec
+from sourced.ml.models.topics import Topics
