+++ sourced/ml/models/topics.py
@@ -1,10 +1,7 @@
-import sys
 from typing import Union
 
-from modelforge import split_strings, assemble_sparse_matrix, \
-    merge_strings, disassemble_sparse_matrix, Model, register_model
-import numpy
-from scipy.sparse import csr_matrix
+from modelforge import register_model, Model, split_strings, assemble_sparse_matrix, merge_strings, \
+    disassemble_sparse_matrix
 
 
 @register_model
@@ -82,38 +79,3 @@ class Topics(Model):
         if not isinstance(labels[0], str):
             raise TypeError("Labels must be strings")
         self._topics = list(labels)
-
-
-def bigartm2asdf_entry(args):
-    """
-    BigARTM "readable" model -> Topics -> Modelforge ASDF.
-    """
-    tokens = []
-    data = []
-    indices = []
-    indptr = [0]
-    if args.input != "-":
-        fin = open(args.input)
-    else:
-        fin = sys.stdin
-    try:
-        # the first line is the header
-        fin.readline()
-        for line in fin:
-            items = line.split(";")
-            tokens.append(items[0])
-            nnz = 0
-            for i, v in enumerate(items[2:]):
-                if v == "0":
-                    continue
-                nnz += 1
-                data.append(float(v))
-                indices.append(i)
-            indptr.append(indptr[-1] + nnz)
-    finally:
-        if args.input != "-":
-            fin.close()
-    data = numpy.array(data, dtype=numpy.float32)
-    indices = numpy.array(indices, dtype=numpy.int32)
-    matrix = csr_matrix((data, indices, indptr), shape=(len(tokens), len(items) - 2)).T
-    Topics().construct(tokens, None, matrix).save(args.output)
